requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2022.1
crispy-bootstrap5==0.7
Django==4.2.4
django-crispy-forms==2.0
django-datetime-widget2==0.9.5
install==1.3.5
Pillow==10.0.0
python-decouple==3.8
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
